Fix: cart page structure
