+++ scss/components/button.module.scss
@@ -63,6 +63,9 @@
 
   height: 30px;
   width: 30px;
+
+  padding: 0;
+  
   &:hover{
     box-shadow: 0 3px 3px 0 $light-gray
   }
